refactor models to subpackages of vo_models
