+++ vo_models/xlink/__init__.py
@@ -2,4 +2,4 @@
 See: https://www.w3.org/TR/xlink11/
 
 Note: Only implements the simple type TypeValue, used in UWS Job models."""
-from vo_models.xml.xlink.xlink import XlinkType
+from vo_models.xlink.xlink import XlinkType
