--- vo_models/xml/__init__.py
@@ -1,1 +1,0 @@
-"""IVOA resource models using pydantic-xml for XML representation."""
